withdraw: Fix json returns
To maintain simplicity decided to create multiple objects with different
names so the serializer just do it's job on the way out. Ideally I'd model
the actual Deposit Transaction, but for simplicity's sake, the transaction
would be very much the same as the account object. So we just duplicate it.

Also fixed returning types because StatusCode() is easier than Created()
and whatnot.